major refactor to move tar and file utilities into central package
