+++ actions/docker_build.go
@@ -11,8 +11,8 @@ import (
 	"time"
 
 	"github.com/arschles/gci/config"
-	"github.com/arschles/gci/dockutil"
 	"github.com/arschles/gci/log"
+	dockutil "github.com/arschles/gci/util/docker"
 	"github.com/codegangsta/cli"
 	docker "github.com/fsouza/go-dockerclient"
 )
