+++ util/docker/attach_opts.go
@@ -1,4 +1,4 @@
-package dockutil
+package docker
 
 import (
 	"io"
