+++ util/docker/attach_wait.go
@@ -1,4 +1,4 @@
-package dockutil
+package docker
 
 import (
 	docker "github.com/fsouza/go-dockerclient"
