+++ util/docker/build_result.go
@@ -1,0 +1,3 @@
+package docker
+
+type BuildResult struct{}
