+++ util/docker/client.go
@@ -1,4 +1,4 @@
-package dockutil
+package docker
 
 import (
 	"os"
