+++ util/docker/constants.go
@@ -1,4 +1,4 @@
-package dockutil
+package docker
 
 const (
 	GolangImage     = "golang:1.5.2"
