+++ util/docker/create_start_opts.go
@@ -1,4 +1,4 @@
-package dockutil
+package docker
 
 import (
 	"fmt"
