+++ util/file/open_stat_test.go
@@ -1,4 +1,4 @@
-package server
+package file
 
 import (
 	"os"
